requirements.txt
python-dotenv~=0.14.0
spotipy~=2.16.0
discord.py~=1.5.0
schedule>=0.6.0